+++ package.json
@@ -22,6 +22,7 @@
     "@tanstack/react-table": "^8.16.0",
     "antd": "^5.19.2",
     "axios": "^1.7.2",
+    "chart.js": "^4.4.4",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "lucide-react": "^0.367.0",
@@ -30,10 +31,12 @@
     "react": "18.2.0",
     "react-apexcharts": "^1.4.1",
     "react-barcode": "^1.5.3",
+    "react-chartjs-2": "^5.2.0",
     "react-dom": "18.2.0",
     "react-hook-form": "^7.53.0",
     "react-loader-spinner": "^6.1.6",
     "react-spinners": "^0.14.1",
+    "sharp": "^0.33.5",
     "sonner": "^1.4.41",
     "tailwind-merge": "^2.5.2",
     "tailwindcss-animate": "^1.0.7",
